Address most comments
