+++ docs/source/references.bib
@@ -16,6 +16,19 @@
       primaryClass={quant-ph},
       url={https://arxiv.org/abs/2502.16404},
 }
+@article{Garcia_2024_01663,
+  title = {On the Hardness of Measuring Magic},
+  author = {Garcia, Roy J. and Bhole, Gaurav and Bu, Kaifeng and Chen, Liyuan and Arthanari, Haribabu and Jaffe, Arthur},
+  journal = {Phys. Rev. Research},
+  volume = {7},
+  pages = {033271},
+  year = {2025},
+  doi = {10.1103/PhysRevResearch.7.033271},
+  eprint = {2408.01663},
+  archivePrefix = {arXiv},
+  primaryClass = {quant-ph},
+  url = {https://arxiv.org/abs/2408.01663},
+}
 @article{Smith_2025,
   title = {Optimally Generating $\mathfrak{su}({2}^{N})$ Using Pauli Strings},
   author = {Smith, Isaac D. and Cautr\`es, Maxime and Stephen, David T. and Poulsen Nautrup, Hendrik},
