+++ docs/source/user/commutator_graph.rst
@@ -55,27 +55,39 @@ which outputs:
 
 As we can see, two initially commuting operators may fail to commute after Heisenberg-evolution under some dynamics.
 
-Fixed unitary, uniform Paulis, and Pauli instability
------------------------------------------------------
+Pauli instability and ``average_otoc``
+--------------------------------------
 
-The :func:`~paulie.application.otoc.average_otoc` function above averages the OTOC over **Haar-random** evolution :math:`U` drawn from the Lie group generated by a Pauli-string DLA (see the commutator-graph discussion and :mod:`paulie.application.otoc`). That is a different ensemble from the following **fixed-unitary** construction.
+The preceding ``average_otoc`` example forms an expectation by drawing **Haar-random** dynamics :math:`U`
+from the Lie group generated by your DLA. **Pauli instability** (Definition~1 in :cite:t:`Garcia_2024_01663`)
+instead fixes a **chosen** unitary :math:`U` and forms an expectation over **independent uniform** Pauli
+strings :math:`P_1,P_2` on :math:`\{I,X,Y,Z\}^{\otimes n}` (and the OTOC built from those). Both are
+averages over random inputs, but the randomness runs over **different** degrees of freedom: :math:`U` vs
+:math:`(P_1,P_2)`.
 
-For a chosen unitary :math:`U` on :math:`n` qubits (:math:`d = 2^n`) and independent **uniform** Pauli strings :math:`P_1, P_2` on :math:`\{I,X,Y,Z\}^{\otimes n}`, define
+There, ``pauli_instability`` quantifies **magic** (non-Clifford resource) via OTOCs. Corollary~1 of that work highlights when such a quantity can be **measured efficiently** (Pauli sample complexity polynomial in :math:`n`) versus when accurate estimation requires resources exponential in system size — a useful lens when you interpret Monte Carlo estimates from ``mean_abs_otoc_uniform`` or ``pauli_instability``.
 
-.. math::
+Definitions (Definition~1), conventions, and implementation details for those helpers live in the API docs for :mod:`paulie.application.otoc`. Below we only **call** ``pauli_instability`` on prototypical gates.
 
-    \mathrm{OTOC}(U, P_1, P_2) = \frac{1}{d}\,\mathrm{Tr}\!\left(
-    U^\dagger P_1 U P_2 U^\dagger P_1 U P_2\right).
+.. code-block:: python
 
-The function :func:`~paulie.application.pauli_instability.otoc_fixed_unitary` evaluates this quantity for explicit :math:`P_1,P_2`. The **mean absolute OTOC** over the uniform product measure on :math:`(P_1,P_2)` is computed by :func:`~paulie.application.pauli_instability.mean_abs_otoc_uniform`, and **Pauli instability** (Definition 1) is
+    import numpy as np
+    from paulie import pauli_instability
 
-.. math::
+    # Single-qubit Hadamard is Clifford → Pauli instability vanishes (faithfulness).
+    h = np.array([[1, 1], [1, -1]], dtype=np.complex128) / np.sqrt(2)
+    print("Pauli instability of H:", round(pauli_instability(h, method="exact"), 6))
 
-    I(U) = -\log \mathbb{E}_{P_1,P_2}\bigl[\lvert\mathrm{OTOC}(U,P_1,P_2)\rvert\bigr],
+    # Single-qubit T gate: I(T) = log(4/3) (scaling with T gates, k=1 in the paper).
+    t = np.diag([1.0, np.exp(1j * np.pi / 4)]).astype(np.complex128)
+    print("Pauli instability of T:", round(pauli_instability(t, method="exact"), 6))
+    print("log(4/3)             :", round(np.log(4.0 / 3.0), 6))
 
-via :func:`~paulie.application.pauli_instability.pauli_instability`.
+Typical output:
 
-**Exact vs Monte Carlo.** Summing over all :math:`16^n` Pauli pairs is implemented as ``method="exact"`` but is **only allowed for at most four qubits** to avoid accidental huge loops. For more qubits, use ``method="monte_carlo"`` with ``num_samples`` as needed.
+.. code-block:: bash
 
-**Randomness.** Monte Carlo draws use :class:`numpy.random.Generator`. Pass an integer ``seed`` to :func:`~paulie.application.pauli_instability.mean_abs_otoc_uniform` or :func:`~paulie.application.pauli_instability.pauli_instability` for reproducible estimates; with ``seed=None`` each call uses a new default generator.
+    Pauli instability of H: 0.0
+    Pauli instability of T: 0.287682
+    log(4/3)             : 0.287682
